+++ apps/worker/src/app/workflow/usecases/send-message/send-message.usecase.ts
@@ -30,7 +30,7 @@ import {
 import { ContextResolved, ExecuteOutput } from '@novu/framework/internal';
 import {
   DeliveryLifecycleDetail,
-  DeliveryLifecycleStatus,
+  DeliveryLifecycleStatusEnum,
   DigestTypeEnum,
   ExecutionDetailsSourceEnum,
   ExecutionDetailsStatusEnum,
@@ -138,7 +138,7 @@ export class SendMessage {
       return {
         status: SendMessageStatus.SKIPPED,
         deliveryLifecycleState: {
-          status: DeliveryLifecycleStatus.SKIPPED,
+          status: DeliveryLifecycleStatusEnum.SKIPPED,
           detail: !channelPreference.result
             ? DeliveryLifecycleDetail.SUBSCRIBER_PREFERENCE
             : DeliveryLifecycleDetail.USER_STEP_CONDITION,