+++ apps/worker/src/app/workflow/usecases/update-job-status/set-job-as.command.ts
@@ -1,5 +1,5 @@
 import { EnvironmentLevelWithUserCommand } from '@novu/application-generic';
-import { IsDefined } from 'class-validator';
+import { IsDefined, IsOptional } from 'class-validator';
 
 export class SetJobAsCommand extends EnvironmentLevelWithUserCommand {
   @IsDefined()
@@ -9,4 +9,7 @@ export class SetJobAsCommand extends EnvironmentLevelWithUserCommand {
 export class SetJobAsFailedCommand extends SetJobAsCommand {
   @IsDefined()
   organizationId: string;
+
+  @IsOptional()
+  isLastJobInWorkflow?: boolean;
 }